+++ src/ninja/mbedded/ninjaterm/util/Decoding/Decoder.java
@@ -1,5 +1,9 @@
 package ninja.mbedded.ninjaterm.util.Decoding;
 
+import javafx.beans.property.ReadOnlyObjectProperty;
+import javafx.beans.property.ReadOnlyObjectWrapper;
+import javafx.beans.property.SimpleObjectProperty;
+
 import java.io.UnsupportedEncodingException;
 
 /**
@@ -9,11 +13,11 @@ import java.io.UnsupportedEncodingException;
  *
  * @author          Geoffrey Hunter <gbmhunter@gmail.com> (www.mbedded.ninja)
  * @since           2016-08-25
- * @last-modified   2016-08-25
+ * @last-modified   2016-10-07
  */
 public class Decoder {
 
-    public DecodingOptions decodingOption = DecodingOptions.ASCII;
+    public SimpleObjectProperty<DecodingOptions> decodingOption = new SimpleObjectProperty<>(DecodingOptions.ASCII);
 
     public Decoder() {
 
@@ -22,22 +26,19 @@ public class Decoder {
     public String parse(byte[] data) {
 
         String output;
-        if(decodingOption == DecodingOptions.ASCII) {
+        if(decodingOption.get() == DecodingOptions.ASCII) {
             try {
                 output = new String(data, "UTF-8");
             } catch (UnsupportedEncodingException e) {
                 throw new RuntimeException(e);
             }
-        } else if(decodingOption == DecodingOptions.HEX) {
+        } else if(decodingOption.get() == DecodingOptions.HEX) {
             output = BytesToString.bytesToHex(data);
         } else {
             throw new RuntimeException("formatting option was not recognised.");
         }
 
         return output;
-
     }
-
-
 
 }